# school-communication-platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Communication Platform API
  description: >
    REST API for five connected dashboards (Super Admin, Admin, Teacher,
    Student, Parent). Messaging, document sharing, announcements and
    notifications. Full reference: docs/API.md and /docs/api.html.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /documents:
    get:
      summary: Documents I can access
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: folderId, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents: { type: array }
    post:
      summary: Upload a document (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                description: { type: string }
                share: { type: string, description: 'JSON array of {targetType, targetId}' }
      responses:
        '201': { description: Uploaded }
        '400': { description: Invalid file }
        '413': { description: Too large }
...
  /parents/documents:
    get:
      summary: Documents shared with the parent or their children
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: classId, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Success
...
  /announcements:
    get:
      summary: Announcements visible to me
      responses:
        '200':
          description: Success
    post:
      summary: Publish an announcement (admin/super admin/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, targetType]
              properties:
                title: { type: string }
                content: { type: string }
                targetType: { type: string, enum: [all, role, class, parents_of_class, staff, students, users] }
                targetValue: { type: string }
                important: { type: boolean }
      responses:
        '201': { description: Published }
...